+++ src/core/FunctionOfEpoch/ObjectState.js
@@ -1,5 +1,5 @@
 import FunctionOfEpochAbstract from "./Abstract";
-import { sim } from "../Simulation";
+import { sim } from "../simulation-engine";
 
 export default class FunctionOfEpochObjectState extends FunctionOfEpochAbstract
 {
