+++ src/core/Trajectory/Abstract.js
@@ -7,7 +7,7 @@ import ReferenceFrameInertialDynamic from "../ReferenceFrame/InertialDynamic";
 import KeplerianEditor from "../KeplerianEditor";
 import {Quaternion} from "../algebra";
 import ReferenceFrameAbstract from "../ReferenceFrame/Abstract";
-import { sim } from "../Simulation";
+import { sim } from "../simulation-engine";
 import Events from "../Events";
 import FlightEventSOIDeparture from "../../modules/PatchedConics/FlightEvent/SOIDeparture";
 import FlightEventSOIArrival from "../../modules/PatchedConics/FlightEvent/SOIArrival";
