+++ src/core/simulation-engine.js
@@ -239,4 +239,5 @@ export let sim = null;
 
 export function createSimulationEngine(simulationModel) {
     sim = new SimulationEngine(simulationModel)
+    return sim;
 }
