+++ src/core/visual/Angle.js
@@ -6,7 +6,7 @@ import {RF_BASE} from "../ReferenceFrame/Factory";
 import VirtualPlane from "./VirtualPlane";
 import {TWO_PI, rad2deg, deg2rad} from "../algebra";
 import ArrowObject from "./Arrow";
-import { sim } from "../Simulation";
+import { sim } from "../simulation-engine";
 
 export default class VisualAngle extends VisualModelAbstract
 {
