+++ src/core/visual/Label.js
@@ -1,7 +1,7 @@
 import * as THREE from "three";
 
 import VisualModelAbstract from "./ModelAbstract";
-import { sim } from "../Simulation";
+import { sim } from "../simulation-engine";
 
 export default class VisualLabel extends VisualModelAbstract {
     constructor(positionOfEpoch, parameters) {
