+++ src/core/visual/Raycaster.js
@@ -1,7 +1,7 @@
 import * as THREE from "three";
 
 import {deg2rad} from "../algebra";
-import { sim } from "../Simulation";
+import { sim } from "../simulation-engine";
 
 export default class VisualRaycaster
 {
