+++ src/core/visual/SpriteStatic.js
@@ -1,6 +1,6 @@
 
 import VisualSprite from "./Sprite";
-import { sim } from '../Simulation';
+import { sim } from '../simulation-engine';
 
 export default class VisualSpriteStatic extends VisualSprite
 {
