+++ src/interface/StarSystemEntry.js
@@ -1,4 +1,4 @@
-import { sim } from "../core/Simulation";
+import { sim } from "../core/simulation-engine";
 
 export default class StarSystemEntry
 {
