+++ src/modules/KSP/ModuleKSP.js
@@ -1,5 +1,5 @@
 import Module from "../../core/Module";
-import { sim } from "../../core/Simulation";
+import { sim } from "../../core/simulation-engine";
 
 export default class ModuleKSP extends Module
 {
