+++ src/modules/SolarSystem/Trajectory/ELP2000.js
@@ -4,7 +4,7 @@ import {ReferenceFrame} from "../../../core/ReferenceFrame/Factory";
 import {deg2rad, Vector} from "../../../core/algebra";
 import StateVector from "../../../core/StateVector";
 import ModuleSolarSystem from "../ModuleSolarSystem";
-import { sim } from "../../../core/Simulation";
+import { sim } from "../../../core/simulation-engine";
 
 export default class TrajectoryELP2000 extends TrajectoryKeplerianAbstract
 {
